+++ K8sUtils/public/Invoke-HelmUpgrade.ps1
@@ -50,6 +50,9 @@ If set, don't actually do the helm upgrade
 .PARAMETER ColorType
 How to colorize the output. Defaults to DevOps if TF_BUILD env var, otherwise ANSI colors
 
+.PARAMETER LogFileFolder
+If specified, pod logs will be written to this folder
+
 .EXAMPLE
     $parms = "preHook.fail=$HookFail," +
               "preHook.imageTag=$HookTag," +
@@ -130,62 +133,14 @@ function Invoke-HelmUpgrade {
         [switch] $SkipRollbackOnError,
         [switch] $DryRun,
         [ValidateSet("None","ANSI","DevOps")]
-        [string] $ColorType = $script:ColorType
+        [string] $ColorType = $script:ColorType,
+        [string] $LogFileFolder
     )
     Set-StrictMode -Version Latest
     $ErrorActionPreference = "Stop"
 
     $minPreHookTimeoutSecs = 120
     $minPodTimeoutSecs = 180
-
-    function rollbackAndWarn {
-        [CmdletBinding()]
-        param ($SkipRollbackOnError, $releaseName, $msg, $prevVersion)
-
-        try {
-            Write-Verbose "helm status --namespace $Namespace $ReleaseName -o json"
-            $currentReleaseVersion = helm status --namespace $Namespace $ReleaseName -o json | ConvertFrom-Json -Depth 10 -AsHashtable # AsHashTable allows for duplicate keys in env, etc.
-            if (!$currentReleaseVersion -or !($currentReleaseVersion.ContainsKey('version'))) {
-                Write-Status "Unexpected response from helm status, not rolling back" -LogLevel warning
-                Write-Status "Current helm release: $($currentReleaseVersion | ConvertTo-Json -Depth 20 -EnumsAsStrings)"
-                return [RollbackStatus]::HelmStatusFailed
-            }
-            Write-Verbose "Current version of $ReleaseName is $($currentReleaseVersion.version)"
-            if (!$currentReleaseVersion -or $currentReleaseVersion.version -eq $prevVersion) {
-                Write-Status "No change in release '$ReleaseName', not rolling back" -LogLevel warning
-                # throw "$msg, no change"
-                Write-Warning "$msg, no change"
-                return [RollbackStatus]::NoChange
-            }
-
-            if (!$SkipRollbackOnError) {
-                Write-Header "Rolling back release '$ReleaseName' due to errors" -LogLevel Error
-                $errFile = Get-TempLogFile
-                helm rollback $ReleaseName --wait 2>&1 | Tee-Object $errFile | Write-MyHost
-                $exit = $LASTEXITCODE
-                $content = Get-Content $errFile -Raw
-                if ($exit -ne 0 -and ($content -like '*Error: release has no 0 version*' -or $content -like '*Error: release: not found*')) {
-                    Write-Verbose "Last exit code on rollback was $exit."
-                    Write-Status "helm rollback failed, trying uninstall" -LogLevel Error -Char '-'
-                    helm uninstall $ReleaseName 2>&1 | Write-MyHost
-                }
-                Remove-Item $errFile -ErrorAction SilentlyContinue
-                Write-Footer "End rolling back release '$ReleaseName' due to errors"
-                # throw "$msg, rolled back"
-                Write-Warning "$msg, rolled back"
-                return [RollbackStatus]::RolledBack
-            } else {
-                # throw "$msg, but not rolling back since -SkipRollbackOnError was specified"
-                Write-Warning "$msg, but not rolling back since -SkipRollbackOnError was specified"
-                return [RollbackStatus]::Skipped
-            }
-            return [RollbackStatus]::DeployedOk
-        } catch {
-            Write-Warning "Caught error rolling back in catch"
-            Write-Error "$_`n$($_.ScriptStackTrace)"
-            return [RollbackStatus]::HelmStatusFailed
-        }
-    }
 
     if (!(Get-Command helm -ErrorAction SilentlyContinue) -or !(Get-Command kubectl -ErrorAction SilentlyContinue)) {
         throw "helm and kubectl must be installed and in the path"
@@ -280,7 +235,8 @@ function Invoke-HelmUpgrade {
                                                         -Namespace $Namespace `
                                                         -TimeoutSec 1 `
                                                         -PollIntervalSec $PollIntervalSec `
-                                                        -PodType PreInstallJob
+                                                        -PodType PreInstallJob `
+                                                        -LogFileFolder $LogFileFolder
             Write-Verbose "Prehook status is $($hookStatus | ConvertTo-Json -Depth 5 -EnumsAsStrings)"
             $status.PreHookStatus = $hookStatus
         }
@@ -305,7 +261,8 @@ function Invoke-HelmUpgrade {
             $podStatuses = Get-DeploymentStatus -TimeoutSec $PodTimeoutSecs `
                                     -Namespace $Namespace `
                                     -Selector $DeploymentSelector `
-                                    -PollIntervalSec $PollIntervalSec
+                                    -PollIntervalSec $PollIntervalSec `
+                                    -LogFileFolder $LogFileFolder
 
             $status.PodStatuses = @() # ?? can't assign the array to podStatuses
             Write-Verbose "Pod statuses are $($podStatuses | ConvertTo-Json -Depth 5 -EnumsAsStrings)"